back/requirements.txt
asgiref==3.3.4
astroid==2.5.3
colorama==0.4.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.0.3
Pillow==8.2.0
PyJWT==1.7.1
pylint==2.7.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
